Fix Download exception when kodi isn't playing
* Fix the error exception from kodi when downloading subtitle is done
when kodi isn't playing (Added an if conditition in mirror_sub function)
* Note: When the user will press on a subtitle, the subtitle will still
be downloaded to the temp folder in the addon folder in userdata (As it
done currently for normal downloading when kodi is playing), after that
the search subtitles dialog will be closed. But this is not should be
relevant to the user for the main purpose of using it only for watching
the subtitles result list.

In Addition:
* Set a new local variable __isKodiPlaying__ instead of writing
xbmc.Player().isPlaying() on each place.
* Fix speeling "Serach for" to "Search for..."