Critical Fix - If Condition when title is empty.
Salts and Local ontent for example gives empty title from the player,
while exodus and specto gives a title. So in Exodus case it always goes
to the "else" and didn't show results.
The if for kodi is playing must be inside the if for empty title. So
when the title is empty it takes the right title. The "else" must be
inside because 2 addons behave different.

I check this and now it's ok.