Critical Fix - Regarding If condition when title is empty
Salts and Local content for example gives empty title from the player,
while exodus and specto gives a title. So currently in Exodus case it
always goes to the "else" with the "fake" title and didn't show results.
When with salts it go into the right place and fix the title.

The fix:
Basically, put the "fake" title few lines before, in the first condition
if kodi is playing (Instead of the empty title as it now), and bring
back the if condition when title is empty as it was before (without
checking there if kodi is playing). But I didn't check this like that.

I check this and now it's ok.